archive/lyria-music-player/requirements.txt
google-genai>=1.16.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
